requirements.txt
numpy<2
Pillow
pandas
scikit-learn
joblib
lightgbm
shap
matplotlib
seaborn
tensorflow
tqdm
pyarrow
